making new getVotingData function to deal with async nightmare
relates to #79
