+++ src/helper_functions/extractContexts.js
@@ -1,13 +1,13 @@
 const get = require('../database/get_data');
 
-function extractContexts(senderID, intent) {
+function extractContexts(senderID, intent, cb) {
   get.party(senderID, (err, res) => {
     if (err) {
       return err;
     }
     console.log(res + intent);
     const partyKey = res + intent;
-    return partyKey;
+    cb(senderID, intent, partyKey);
   });
 }
