+++ src/helper_functions/findLocalReply.js
@@ -2,6 +2,7 @@ const sendToFB = require('./sendToFB');
 const construct = require('./answer_objects');
 const get = require('./../database/get_data');
 const extractContexts = require('./extractContexts');
+const getVotingData = require('./getVotingData');
 
 function constructLocal(senderID, key, answerObjects) {
   const messageData = {
@@ -19,24 +20,7 @@ function findLocalReply(senderID, intent) {
   // console.log('intent is', intent);
 
   if (intent === 'brexit' || intent === 'tuitionFees' || intent === 'syria') {
-    const partyKey = extractContexts(senderID, intent);
-    console.log('partyKey is ', partyKey);
-    get.partyVotes(partyKey, (err, res) => {
-      if (err) {
-        return err;
-      }
-      const partyVotesObj = res.rows[0];
-      console.log('partyVotesObj is ', partyVotesObj);
-      const answerObjects = construct(partyVotesObj, null);
-      console.log('answerObjects are ', answerObjects);
-      for (const key in answerObjects) {
-        if (key === intent) {
-          const messageData = constructLocal(senderID, key, answerObjects);
-          sendToFB(messageData);
-          // boolean = true;
-        }
-      }
-    });
+    extractContexts(senderID, intent, getVotingData);
   } else {
     get.firstName(senderID, (err, firstName) => {
       if (err) {
